python games/requirements.txt
git-filter-repo==2.34.0
ion-Numworks==1.2.3
kandinsky==1.9
keyboard==0.13.5
matplotlib-inline==0.1.6
matplotlib==3.6.0
numpy==1.23.3
pip==22.2.2
pygame==2.1.2
future~=0.18.2